Working visualization for the discrete simulator. Issues: Currently the turtles do not accurately represent the heading of their direction (they don't point where they're going and they don't point in the direction they were going if you go back in time). This could be fixed by using turtle.headingto(location) which will point the turtle in the direction of a coordinate, turtle.setpos() only moves a turtle and does not orient the turtle to the coordinate it is being moved to.
Despite being explicitly told to have color, all l turtles have turned back to black for some reason after initially having color. This error was caught later and the exact cause is not known yet.

There is an issue with the turtle graphics window freezing for 5 seconds at the start of the simulation where it cannot be interacted with.

A coordinate conversion function is needed to be able to map the turtle to the background diagram, if this could be solved by sending the turtle origin to the diagram origin, the code could be much simpler.

Turtle.clear() within coordinate_crawl method needs to be fixed as it accidentally clears text of turtles texts that are attached to turtle trucks that are not moving. Turtle trucks that have finished end up being displayed with no name.

All turtles could be enlarged for visibility.

Turtle name text sometimes overlaps.

Visualizer does not work on other devices, need to fix the root path for turtle.scrbackground() so it runs without needing to change the path of the road network diagram.